# docs/swagger.yaml
basePath: /
host: localhost:3000
info:
  contact: {}
  description: P2P file sharing with content-addressable storage, collection management,
    versioning, merge/fork/pull.
  title: Peerdrive API
  version: "1.0"
paths:
...
  /actions/merge:
    post:
      consumes:
      - application/json
      description: 'Merge entries from source collection into local. Strategy: "ours"
        keeps local, "theirs" accepts source, "manual" returns conflicts list.'
      parameters:
      - description: Merge request
        in: body
        name: body
        required: true
        schema:
          properties:
            collection_name:
              type: string
            source_coll_name:
              type: string
            source_username:
              type: string
            strategy:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: merge complete
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Collection not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: conflicts list (when strategy=manual)
          schema:
            additionalProperties: true
            type: object
      summary: Merge a source collection into the local collection
      tags:
      - actions
...
  /tasks/{id}:
    get:
      description: Query the status and result of an async task (e.g., pull)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: task details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid task ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get task status by ID
      tags:
      - tasks
swagger: "2.0"